+++ generators/app/constants.js
@@ -1,3 +1,7 @@
+const { filesGraphql } = require('./filesGraphql'),
+  { filesExpress } = require('./filesExpessjs'),
+  { filesCommon } = require('./filesCommon');
+
 exports.NODE_DEFAULT_VERSION = '10.14.1';
 exports.NPM_DEFAULT_VERSION = '6.4.1';
 exports.ORM_OPTIONS = ['sequelize', 'mongoose'];
@@ -11,10 +15,25 @@ exports.DEPLOY_STRATEGIES = ['aws', 'heroku'];
 exports.OPTIONALS_FEATURES = ['coveralls', 'rollbar', 'cors', 'helmet'];
 exports.CI_OPTIONS = ['jenkins', 'travis'];
 exports.TESTING_OPTIONS = ['mocha-chai', 'jest-supertest'];
+exports.TECHNOLOGY_OPTIONS = ['expressJS', 'graphQL'];
 
 exports.TRAINING_CONFIG = {
   projectName: 'WTraining',
   projectDescription: 'WTraining',
+  nodeVersion: exports.NODE_DEFAULT_VERSION,
+  npmVersion: exports.NPM_DEFAULT_VERSION,
+  database: true,
+  orm: { sequelize: true },
+  sequelizeVersion: exports.SEQUELIZE_DEFAULT_VERSION,
+  sequelizeDialect: exports.SEQUELIZE_DEFAULT_DIALECT,
+  deployStrategy: { heroku: true },
+  optionalsFeatures: {},
+  ci: 'travis',
+  testing: 'jest-supertest'
+};
+exports.TRAINING_GRAPHQL_CONFIG = {
+  projectName: 'WTrainingGraphql',
+  projectDescription: 'WTrainingGraphql',
   nodeVersion: exports.NODE_DEFAULT_VERSION,
   npmVersion: exports.NPM_DEFAULT_VERSION,
   database: true,
@@ -37,172 +56,4 @@ exports.TUTORIALS = {
   NPM: 'https://github.com/creationix/nvm#install-script'
 };
 
-exports.files = [
-  {
-    name: 'Jenkinsfile',
-    condition: answers => answers.ci === 'jenkins'
-  },
-  {
-    directory: '.woloxci',
-    name: 'config.yml',
-    condition: answers => answers.ci === 'jenkins'
-  },
-  {
-    directory: '.woloxci',
-    name: 'Dockerfile',
-    condition: answers => answers.ci === 'jenkins'
-  },
-  {
-    name: 'Procfile',
-    condition: answers => answers.deployStrategy.heroku
-  },
-  {
-    name: 'Dockerfile',
-    condition: answers => answers.docker
-  },
-  {
-    name: 'Dockerrun.aws.json',
-    condition: answers => answers.docker && answers.deployStrategy.aws
-  },
-  {
-    name: '.travis.yml',
-    condition: answers => answers.ci === 'travis'
-  },
-  {
-    name: '.sequelizerc',
-    condition: answers => answers.orm && answers.orm.sequelize
-  },
-  {
-    directory: 'migrations',
-    name: 'index.js',
-    condition: answers => answers.orm && answers.orm.sequelize
-  },
-  {
-    directory: 'migrations/migrations',
-    name: '.keep',
-    condition: answers => answers.orm && answers.orm.sequelize
-  },
-  {
-    directory: 'config',
-    name: 'db.ejs',
-    condition: answers => answers.orm && (answers.orm.sequelize || answers.orm.mongoose)
-  },
-  {
-    directory: 'app/models',
-    name: 'index.js',
-    condition: answers => answers.orm && answers.orm.sequelize
-  },
-  {
-    directory: 'test/factory',
-    name: 'factory_by_models.ejs',
-    condition: answers => answers.orm && answers.orm.sequelize && answers.testing === 'jest-supertest'
-  },
-  {
-    directory: 'test',
-    name: 'setup.js',
-    condition: answers => answers.orm && answers.orm.sequelize && answers.testing === 'jest-supertest'
-  },
-  {
-    directory: 'test',
-    name: 'app.spec.ejs',
-    condition: answers => answers.testing === 'mocha-chai'
-  },
-  {
-    name: 'README.md'
-  },
-  {
-    name: 'pull_request_template.md'
-  },
-  {
-    name: 'package.json'
-  },
-  {
-    name: 'LICENSE.md'
-  },
-  {
-    name: 'console.ejs'
-  },
-  {
-    name: 'app.ejs'
-  },
-  {
-    name: 'server.ejs'
-  },
-  {
-    directory: 'documentation',
-    name: 'index.js'
-  },
-  {
-    directory: 'documentation/schemas',
-    name: 'index.js'
-  },
-  {
-    directory: 'documentation/schemas',
-    name: 'user.js'
-  },
-  {
-    directory: 'documentation/paths',
-    name: 'index.js'
-  },
-  {
-    directory: 'documentation/paths',
-    name: 'user.js'
-  },
-  {
-    name: '.nvmrc'
-  },
-  {
-    name: 'gitignore',
-    newName: '.gitignore'
-  },
-  {
-    name: '.eslintrc.js'
-  },
-  {
-    name: '.eslintignore'
-  },
-  {
-    directory: 'config',
-    name: 'development.ejs'
-  },
-  {
-    directory: 'config',
-    name: 'production.ejs'
-  },
-  {
-    directory: 'config',
-    name: 'testing.ejs'
-  },
-  {
-    directory: 'config',
-    name: 'index.ejs'
-  },
-  {
-    directory: 'app/controllers',
-    name: 'healthCheck.js'
-  },
-  {
-    directory: 'app/services',
-    name: '.keep'
-  },
-  {
-    directory: 'app',
-    name: 'errors.js'
-  },
-  {
-    directory: 'app',
-    name: 'routes.js'
-  },
-  {
-    directory: 'app/middlewares',
-    name: 'errors.js'
-  },
-  {
-    directory: 'app/logger',
-    name: 'index.js'
-  },
-  {
-    directory: 'app/middlewares',
-    name: 'apiInfo.js'
-  }
-];
+exports.files = [...filesCommon, ...filesExpress, ...filesGraphql];